requirements.txt
langchain
langchain-community
langchain-openai
chromadb
openai
python-docx
python-dotenv
tiktoken
